sprint8/backend/requirements.txt
# FastAPI + server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Document processing
pymupdf==1.24.3          # PyMuPDF — fast PDF text extraction
pytesseract==0.3.10      # OCR fallback for scanned PDFs
pdf2image==1.17.0        # Convert PDF pages to images for OCR
python-docx==1.1.2       # DOCX text extraction

# AI / Embeddings
sentence-transformers==3.0.1   # Local embeddings (free, no API needed)
torch==2.3.0                   # Required by sentence-transformers

# Vector database
pinecone-client==4.1.0   # Pinecone vector store

# LangChain for chunking
langchain==0.2.6
langchain-text-splitters==0.2.2

# Task queue
celery==5.4.0
redis==5.0.6             # Celery broker

# Database
supabase==2.5.0

# AWS S3 (for document storage)
boto3==1.34.131

# Utilities
pydantic==2.7.4
pydantic-settings==2.3.3
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.3.0          # Retry logic
